# docs/rfq_manager_ms_openapi_v1.yaml
openapi: 3.0.3
info:
  title: rfq_manager_ms — RFQ Lifecycle Management API
  description: |
    Interface contract for the RFQ Manager microservice (V1).
    Front-to-back design: 9 UI screens → 47 user actions → 6 resources → 28 endpoints → 9 tables.
  version: "1.0.0"
  contact:
    name: RFQ Manager Team
servers:
  - url: /rfq-manager/v1
    description: Base path — convention /{ms_name}/{version}
...
security:
  - bearerAuth: []

# ═══════════════════════════════════════════════════
# PATHS
# ═══════════════════════════════════════════════════
paths:
...
  /rfqs:
    post:
      tags: [RFQ]
      summary: "#1 — Create RFQ"
      description: |
        Creates a new RFQ and auto-generates all rfq_stage instances from the selected workflow.
        Back-calculates planned_start / planned_end from deadline.
        Sets current_stage_id = first stage, status = "In preparation".
        Note: Draft status is reserved for a future "save without submitting" flow — this endpoint always creates in "In preparation".
      operationId: createRfq
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RfqCreateRequest'
      responses:
        '201':
          description: RFQ created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfqDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    get:
      tags: [RFQ]
      summary: "#2 — List RFQs"
      description: |
        Paginated list with search, status filter, priority filter, and sort.
        Used by index.html portfolio table and workspace.html RFQ selector.
        Excludes Draft RFQs by default — pass ?status=Draft to include them.
      operationId: listRfqs
      parameters:
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/StatusFilterParam'
        - $ref: '#/components/parameters/PriorityFilterParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated RFQ list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfqListResponse'
  /rfqs/{rfqId}:
    get:
      tags: [RFQ]
      summary: "#3 — Get RFQ detail"
      description: Full RFQ detail — all list fields + industry, priority, description, workflow_id, current_stage_id, timestamps.
      operationId: getRfq
      parameters:
        - $ref: '#/components/parameters/RfqIdParam'
      responses:
        '200':
          description: RFQ detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfqDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [RFQ]
      summary: "#4 — Update RFQ"
      description: |
        Update RFQ metadata or set final outcome (Awarded / Lost / Cancelled).
        Changing deadline triggers recalculation of all rfq_stage planned dates.
      operationId: updateRfq
      parameters:
        - $ref: '#/components/parameters/RfqIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RfqUpdateRequest'
      responses:
        '200':
          description: RFQ updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfqDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /rfqs/{rfqId}/stages:
    get:
      tags: [RFQ_Stage]
      summary: "#10 — List stages for RFQ"
      description: All stages for an RFQ, ordered by `order`. Used by rfq_overview.html timeline and workspace.html stage tabs.
      operationId: listRfqStages
      parameters:
        - $ref: '#/components/parameters/RfqIdParam'
      responses:
        '200':
          description: Stage list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RfqStage'
  /rfqs/{rfqId}/stages/{stageId}:
    get:
      tags: [RFQ_Stage]
      summary: "#11 — Get stage detail"
      description: Same as list item + embedded subtasks[], files[], full captured_data, blocker_status, blocker_reason_code.
      operationId: getRfqStage
      parameters:
        - $ref: '#/components/parameters/RfqIdParam'
        - $ref: '#/components/parameters/StageIdParam'
      responses:
        '200':
          description: Stage detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfqStageDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [RFQ_Stage]
      summary: "#12 — Update stage"
      description: |
        Update progress, captured_data, blocker_status, blocker_reason_code, assigned_team.
        Set blocker_status = "Blocked" with a reason code to flag a blocker.
        Set blocker_status = "Resolved" to clear it.
      operationId: updateRfqStage
      parameters:
        - $ref: '#/components/parameters/RfqIdParam'
        - $ref: '#/components/parameters/StageIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RfqStageUpdateRequest'
      responses:
        '200':
          description: Stage updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfqStageDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /rfqs/{rfqId}/stages/{stageId}/notes:
    post:
      tags: [RFQ_Stage]
      summary: "#13 — Add note to stage"
      description: |
        Append-only — prevents concurrent overwrite.
        Auto-sets: user_name (from auth), created_at (server timestamp), id.
      operationId: addStageNote
      parameters:
        - $ref: '#/components/parameters/RfqIdParam'
        - $ref: '#/components/parameters/StageIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  example: "Main vessel shell/head plates identified."
      responses:
        '201':
          description: Note added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StageNote'
  /rfqs/{rfqId}/stages/{stageId}/files:
    post:
      tags: [RFQ_Stage]
      summary: "#14 — Upload file to stage"
      description: |
        Multipart upload. Type: "Client RFQ", "Design report", "BOQ / BOM", "Estimation Workbook", "Other".
        V2: type = "Estimation Workbook" will trigger intelligence parsing.
      operationId: uploadStageFile
      parameters:
        - $ref: '#/components/parameters/RfqIdParam'
        - $ref: '#/components/parameters/StageIdParam'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, type]
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                  enum: [Client RFQ, Design report, BOQ / BOM, Estimation Workbook, Other]
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StageFile'

    get:
      tags: [RFQ_Stage]
      summary: "#26 — List files for stage"
      description: Returns list of files for this stage. Also embedded in GET stage detail, but useful for lazy-loading file lists.
      operationId: listStageFiles
      parameters:
        - $ref: '#/components/parameters/RfqIdParam'
        - $ref: '#/components/parameters/StageIdParam'
      responses:
        '200':
          description: File list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StageFile'
  /rfqs/{rfqId}/stages/{stageId}/advance:
    post:
      tags: [RFQ_Stage]
      summary: "#15 — Advance to next stage"
      description: |
        Sets actual_end on current stage → status = Completed.
        Sets actual_start on next stage → status = In Progress.
        Updates RFQ.current_stage_id + recalculates RFQ.progress.
      operationId: advanceStage
      parameters:
        - $ref: '#/components/parameters/RfqIdParam'
        - $ref: '#/components/parameters/StageIdParam'
      responses:
        '200':
          description: Stage advanced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfqStageDetail'
        '403':
          description: Not assigned to this stage's team
        '409':
          description: Stage is blocked (blocker_status = Blocked)
        '422':
          description: Mandatory fields missing in captured_data

  # ──── SUBTASK ────────────────────────────────────
...
  /files/{fileId}/download:
    get:
      tags: [File]
      summary: "#27 — Download file"
      description: Returns file stream or signed Azure Blob URL. Shortcut pattern — fileId is globally unique UUID.
      operationId: downloadFile
      parameters:
        - $ref: '#/components/parameters/FileIdParam'
      responses:
        '200':
          description: File stream or redirect to signed URL
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
  /files/{fileId}:
    delete:
      tags: [File]
      summary: "#28 — Delete file"
      description: Soft delete — sets deleted_at = NOW(). All GETs filter WHERE deleted_at IS NULL.
      operationId: deleteFile
      parameters:
        - $ref: '#/components/parameters/FileIdParam'
      responses:
        '204':
          description: File soft-deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ──── REMINDER ───────────────────────────────────
  /reminders:
    post:
      tags: [Reminder]
      summary: "#20 — Create reminder"
      description: |
        Auto-sets: created_by, created_at, status = "open", send_count = 0.
        Batch process picks these up for notification.
      operationId: createReminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderCreateRequest'
      responses:
        '201':
          description: Reminder created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'

    get:
      tags: [Reminder]
      summary: "#21 — List reminders"
      description: Filter by user, status, rfq_id. Main feed for tasks.html.
      operationId: listReminders
      parameters:
        - name: user
          in: query
          schema:
            type: string
          example: me
        - name: status
          in: query
          schema:
            type: string
            enum: [open, sent, overdue, resolved]
        - name: rfq_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reminder list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reminder'
...
  /reminders/rules/{ruleId}:
    patch:
      tags: [Reminder]
      summary: "#24 — Toggle reminder rule"
      description: Toggle is_active on/off.
      operationId: updateReminderRule
      parameters:
        - $ref: '#/components/parameters/RuleIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                is_active:
                  type: boolean
      responses:
        '200':
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderRule'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from rfq_iam_ms

  # ──── PARAMETERS ─────────────────────────────────
  parameters:
    RfqIdParam:
      name: rfqId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    StageIdParam:
      name: stageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    FileIdParam:
      name: fileId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RuleIdParam:
      name: ruleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SearchParam:
      name: search
      in: query
      schema:
        type: string
      description: Free-text search on name, client
    StatusFilterParam:
      name: status
      in: query
      schema:
        type: string
        enum: [Draft, In preparation, Submitted, Awarded, Lost, Cancelled]
    PriorityFilterParam:
      name: priority
      in: query
      schema:
        type: string
        enum: [normal, critical]
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: "Field to sort by, prefix with - for desc. e.g. -deadline"

  # ──── RESPONSES ──────────────────────────────────
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Validation error — required fields missing or invalid values
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ──── SCHEMAS ────────────────────────────────────
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Not found"
        message:
          type: string
          example: "RFQ with id 'xxx' does not exist"

    # ── RFQ ──
    RfqCreateRequest:
      type: object
      required: [name, client, deadline, owner, workflow_id]
      properties:
        name:
          type: string
          example: "HP Separator Package – Offshore"
        client:
          type: string
          example: "Saudi Aramco"
        industry:
          type: string
          example: "Oil & Gas – Upstream"
        country:
          type: string
          example: "Saudi Arabia"
        priority:
          type: string
          enum: [normal, critical]
          default: normal
        deadline:
          type: string
          format: date
          example: "2025-01-12"
        owner:
          type: string
          example: "Proposals Team A"
        description:
          type: string
          example: "12x HP Separator for offshore platform..."
        workflow_id:
          type: string
          format: uuid
          example: "wf-ghi-long"
        stage_overrides:
          type: array
          items:
            type: object
            properties:
              stage_template_id:
                type: string
                format: uuid
              assigned_team:
                type: string
    RfqUpdateRequest:
      type: object
      properties:
        name:
          type: string
        client:
          type: string
        industry:
          type: string
        country:
          type: string
        priority:
          type: string
          enum: [normal, critical]
        deadline:
          type: string
          format: date
        owner:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [Draft, In preparation, Submitted, Awarded, Lost, Cancelled]
        outcome_reason:
          type: string
          description: "Covers all outcomes: why awarded, why lost, why cancelled"
          example: "Best technical score"
    RfqSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "HP Separator Package"
        client:
          type: string
          example: "Saudi Aramco"
        status:
          type: string
          enum: [Draft, In preparation, Submitted, Awarded, Lost, Cancelled]
        progress:
          type: integer
          example: 42
        deadline:
          type: string
          format: date
        current_stage_name:
          type: string
          example: "BOQ / BOM preparation"
    RfqDetail:
      allOf:
        - $ref: '#/components/schemas/RfqSummary'
        - type: object
          properties:
            industry:
              type: string
            country:
              type: string
            priority:
              type: string
              enum: [normal, critical]
            owner:
              type: string
            description:
              type: string
            workflow_id:
              type: string
              format: uuid
            current_stage_id:
              type: string
              format: uuid
              nullable: true
              description: "NULL when RFQ is in Draft (no stages yet). Set to first stage on finalization."
            outcome_reason:
              type: string
              nullable: true
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    RfqListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RfqSummary'
        total:
          type: integer
          example: 96
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 20
...
    RfqStage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "RFQ received"
        order:
          type: integer
        assigned_team:
          type: string
        status:
          type: string
          enum: [Not started, In Progress, Completed, Skipped]
        progress:
          type: integer
        planned_start:
          type: string
          format: date
        planned_end:
          type: string
          format: date
        actual_start:
          type: string
          format: date
          nullable: true
        actual_end:
          type: string
          format: date
          nullable: true
        blocker_status:
          type: string
          enum: [Blocked, Resolved]
          nullable: true
          description: "NULL = never blocked. Orthogonal to status."
        blocker_reason_code:
          type: string
          nullable: true
          description: "Examples: supplier_delay, missing_client_docs, awaiting_internal_approval, technical_clarification, workbook_issue, budget_hold, pending_lab_results"
    RfqStageDetail:
      allOf:
        - $ref: '#/components/schemas/RfqStage'
        - type: object
          properties:
            captured_data:
              type: object
              description: "Free-form JSON for stage-specific form fields"
            mandatory_fields:
              type: string
              description: "Snapshot from stage_template — used by /advance guard"
            notes:
              type: array
              items:
                $ref: '#/components/schemas/StageNote'
            files:
              type: array
              items:
                $ref: '#/components/schemas/StageFile'
            subtasks:
              type: array
              items:
                $ref: '#/components/schemas/Subtask'
    RfqStageUpdateRequest:
      type: object
      properties:
        progress:
          type: integer
        assigned_team:
          type: string
        captured_data:
          type: object
        blocker_status:
          type: string
          enum: [Blocked, Resolved]
          nullable: true
        blocker_reason_code:
          type: string
          nullable: true
    StageNote:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_name:
          type: string
          example: "Sales Eng."
        text:
          type: string
        created_at:
          type: string
          format: date-time
    StageFile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
          example: "HP_Separator_BOQ_v2.xlsx"
        file_path:
          type: string
        type:
          type: string
          enum: [Client RFQ, Design report, BOQ / BOM, Estimation Workbook, Other]
        uploaded_by:
          type: string
        size_bytes:
          type: integer
          format: int64
        uploaded_at:
          type: string
          format: date-time

    # ── SUBTASK ──
...
    Subtask:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        assigned_to:
          type: string
          nullable: true
        due_date:
          type: string
          format: date
          nullable: true
        progress:
          type: integer
        status:
          type: string
          enum: [Open, In progress, Done]
        created_at:
          type: string
          format: date-time

    # ── REMINDER ──
    ReminderCreateRequest:
      type: object
      required: [rfq_id, type, message, due_date]
      properties:
        rfq_id:
          type: string
          format: uuid
        rfq_stage_id:
          type: string
          format: uuid
          nullable: true
        type:
          type: string
          enum: [internal, external]
        message:
          type: string
          example: "Follow up on nozzle material confirmation"
        due_date:
          type: string
          format: date
        assigned_to:
          type: string
          example: "Engineering"
    Reminder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rfq_id:
          type: string
          format: uuid
        rfq_stage_id:
          type: string
          format: uuid
          nullable: true
        type:
          type: string
          enum: [internal, external]
        message:
          type: string
        due_date:
          type: string
          format: date
        assigned_to:
          type: string
          nullable: true
        status:
          type: string
          enum: [open, sent, overdue, resolved]
        delay_days:
          type: integer
          description: "Computed at query time: MAX(0, NOW()::date - due_date). NOT stored."
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        last_sent_at:
          type: string
          format: date-time
          nullable: true
          description: "Used by batch to rate-limit re-sends"
        send_count:
          type: integer
          description: "Number of times notification was sent"
...
    ReminderRule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Stage overdue alert"
        description:
          type: string
        scope:
          type: string
          description: "Examples: all_rfqs, critical_only, stage_overdue, blocked_stages, upcoming_deadline"
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time